Add files via upload
Bisecting Min–Max Clustering Algorithm is similar to bisecting k-means clustering algorithm with a slight modification. The Min–Max Clustering Algorithm initially computes the minimum of the input set and then finds a point which is at the greatest distance from the minimum. To obtain a certain number of clusters, the main cluster is divided into two clusters using Min-Max algorithm. A cluster is selected on the basis of SSE, in order to split it further. This process is repeated until desired numbers of clusters are obtained. 
Thus, there is a remarkable improvement in the running time of the Bisecting Min-Max Algorithm when compared with the Bisecting K-Means Algorithm.
The steps to categorize elements using Bisecting Min-Max algorithm is as follows:
1.Compute minimum point and maximum point from the data set.
2.Assign other points depending upon the distance between min point and max point.
3.Once the clusters are formed, Calculate the SSE.
4.Compare SSE, bisect the cluster having more SSE using min max algorithm.
Until cluster become equal to K, repeat from step 3.